backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
redis==5.0.1
PyJWT==2.8.0
bcrypt==4.1.2
razorpay==1.4.1
chromadb==0.4.22
google-generativeai==0.3.2
python-dotenv==1.0.0
